refactor(web): diaryの散在ロジック・型をmodel/に集約
authenticateと同様に、feature直下に散らばっていた daily-summary と
types を model/ にまとめ、components・hooks と並ぶ構成に整える。
