+++ application/packages/web/src/client/features/diary/components/read-list-section.tsx
@@ -3,7 +3,7 @@ import type { ReactNode } from 'react'
 import { AnchorLink } from '@/client/components/ui/navigation/link'
 // barrel 経由だと記事スライスのデータ取得 hook（swr 依存）まで巻き込み、Storybook 環境で読み込みに失敗するため
 import MediaIcon from '@/client/features/article/components/media-icon'
-import type { ReadItem } from '@/client/features/diary/types'
+import type { ReadItem } from '@/client/features/diary/model/types'
 import { toSafeExternalPath } from '@/client/lib/url'
 
 interface Props {
