+++ application/packages/web/src/client/features/diary/model/daily-summary.ts
